requirements.txt
PyPDF2
pandas
langchain
langchain-google-genai
google-generativeai
faiss-cpu
streamlit
gtts
python-dotenv
